+++ src/sdc/api/__init__.py
@@ -5,7 +5,7 @@ from ._generator import Generator, SingleVariableGenerator
 from ._spectralio import SpectralIOBased
 from ._reader import Reader, SpectralIOReader, SpectralIOReaderWithLocaleSupport, add_locale_option, parse_reader
 from ._reader import SampleDataReader
-from ._writer import BatchWriter, StreamWriter, SplittableBatchWriter, SplittableStreamWriter, parse_writer
+from ._writer import BatchWriter, StreamWriter, SplittableBatchWriter, SplittableStreamWriter, SpectralIOWriter, parse_writer
 from ._writer import SampleDataBatchWriter, SampleDataStreamWriter, SplittableSampleDataBatchWriter, SplittableSampleDataStreamWriter
 from ._cleaner import Cleaner, parse_cleaner
 from ._utils import safe_deepcopy, locate_file, strip_suffix, load_function