+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 - wai.spectralio-based readers now instantiate the wai.spectralio reader in the `initialize` method
 - wai.spectralio-based writers now instantiate the wai.spectralio writer in the `initialize` method
-- wai.spectralio-based readers/writers now implement the `SpectralIOBased` mixin for a cleaner class hierarchy
+- introduced `SpectralIOBased`, `SpectralIOReader` and `SpectralIOWriter` mixins to wai.spectralio-based
+  readers/writers for a cleaner class hierarchy
 
 
 0.0.1 (2025-06-27)